Move huisbaasje coordinator to separate module
