+++ homeassistant/components/huisbaasje/sensor.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 import logging
-from typing import Any
 
 from energyflip.const import (
     SOURCE_TYPE_ELECTRICITY,
@@ -31,10 +30,7 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-)
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import (
     DATA_COORDINATOR,
@@ -45,6 +41,7 @@ from .const import (
     SENSOR_TYPE_THIS_WEEK,
     SENSOR_TYPE_THIS_YEAR,
 )
+from .coordinator import EnergyFlipUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -222,9 +219,9 @@ async def async_setup_entry(
     async_add_entities: AddConfigEntryEntitiesCallback,
 ) -> None:
     """Set up the sensor platform."""
-    coordinator: DataUpdateCoordinator[dict[str, dict[str, Any]]] = hass.data[DOMAIN][
-        config_entry.entry_id
-    ][DATA_COORDINATOR]
+    coordinator: EnergyFlipUpdateCoordinator = hass.data[DOMAIN][config_entry.entry_id][
+        DATA_COORDINATOR
+    ]
     user_id = config_entry.data[CONF_ID]
 
     async_add_entities(
@@ -233,9 +230,7 @@ async def async_setup_entry(
     )
 
 
-class EnergyFlipSensor(
-    CoordinatorEntity[DataUpdateCoordinator[dict[str, dict[str, Any]]]], SensorEntity
-):
+class EnergyFlipSensor(CoordinatorEntity[EnergyFlipUpdateCoordinator], SensorEntity):
     """Defines a EnergyFlip sensor."""
 
     entity_description: EnergyFlipSensorEntityDescription
@@ -243,7 +238,7 @@ class EnergyFlipSensor(
 
     def __init__(
         self,
-        coordinator: DataUpdateCoordinator[dict[str, dict[str, Any]]],
+        coordinator: EnergyFlipUpdateCoordinator,
         user_id: str,
         description: EnergyFlipSensorEntityDescription,
     ) -> None:
